refactor: move handlers to agent_issuance
