+++ agent_issuance/Cargo.toml
@@ -17,7 +17,10 @@ identity_credential = { git = "https://github.com/iotaledger/identity.rs.git", d
     "credential",
 ] }
 
-[dev-dependencies]
+# TODO: remove this once handlers.rs is removed.
 axum = "0.6"
+
+
+[dev-dependencies]
 tokio = { version = "1", features = ["full"] }
 serde = { version = "1.0", features = ["derive"]}
