+++ agent_issuance/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod command;
 pub mod error;
 pub mod event;
+pub mod handlers;
 pub mod model;
 pub mod queries;
 pub mod services;
